% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchR_helper.R
\name{addAnyReducedMtrx}
\alias{addAnyReducedMtrx}
\title{Add any reduced-dimensions-like matrix to an ArchR project}
\usage{
addAnyReducedMtrx(ArchRProj, name = "myReduced", matrix = NULL)
}
\arguments{
\item{ArchRProj}{An ArchR project}

\item{name}{Name of output MNN matrix}

\item{matrix}{The matrix to be added to ArchR project}
}
\value{

}
\description{
Add any reduced-dimensions-like matrix to an ArchR project
}
\examples{

}
