% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchR_helper.R
\name{addReducedMNN}
\alias{addReducedMNN}
\title{Add MNN adjusted reduced dimensions to an ArchR project}
\usage{
addReducedMNN(
  ArchRProj,
  corCutOff = 0.5,
  reducedDims = "IterativeLSI",
  name = "reducedMNN",
  scaleDims = T,
  scaleDimsAfter = NA,
  groupBy,
  k = 15,
  dimsToUse,
  ...
)
}
\arguments{
\item{ArchRProj}{An ArchR project}

\item{corCutOff}{Cutoff for correlation between library depth and reduced dimensions}

\item{reducedDims}{Name of reducedDims from ArchR project to use in reducedMNN}

\item{name}{Name of output MNN matrix}

\item{scaleDims}{Whether to let ArchR scale reducedDims}

\item{scaleDimsAfter}{Whether to let ArchR scale output reducedDims}

\item{groupBy}{Variable name in ArchR cellColData to adjust for}

\item{k}{}

\item{dimsToUse}{Dimensions to use in reducedMNN}

\item{...}{Other parameters for reducedMNN}
}
\value{

}
\description{
Add MNN adjusted reduced dimensions to an ArchR project
}
