% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchR_addTopicScoreMatrix.R
\name{addTopicScoreMatrix}
\alias{addTopicScoreMatrix}
\title{Add TopicScoreMatrix to ArrowFiles or an ArchRProject}
\usage{
addTopicScoreMatrix(
  genes = NULL,
  peaks = NULL,
  scoreMat = NULL,
  geneModel = "exp(-abs(x)/5000) + exp(-1)",
  matrixName = "TopicScoreMatrix",
  extendUpstream = c(1000, 1e+05),
  extendDownstream = c(1000, 1e+05),
  geneUpstream = 5000,
  geneDownstream = 0,
  useGeneBoundaries = TRUE,
  useTSS = FALSE,
  extendTSS = FALSE,
  geneScaleFactor = 5,
  excludeChr = c("chrY", "chrM"),
  blacklist = NULL,
  tmpFile = NULL,
  subThreads = 1,
  tstart = NULL,
  logFile = NULL
)
}
\arguments{
\item{genes}{A stranded \code{GRanges} object containing the ranges associated with all gene start and end coordinates.}

\item{peaks}{A \code{GRanges} object containing the peaks with a \code{score} column to calculate gene-level scores.}

\item{scoreMat}{}

\item{geneModel}{A string giving a "gene model function" used for weighting peaks for gene score calculation. This string
should be a function of \code{x}, where \code{x} is the stranded distance from the transcription start site of the gene.}

\item{matrixName}{The name to be used for storage of the gene activity score matrix in the provided \code{ArchRProject} or ArrowFiles.}

\item{extendUpstream}{The minimum and maximum number of basepairs upstream of the transcription start site to consider for gene
activity score calculation.}

\item{extendDownstream}{The minimum and maximum number of basepairs downstream of the transcription start site or transcription termination site
(based on 'useTSS') to consider for gene activity score calculation.}

\item{geneUpstream}{An integer describing the number of bp upstream the gene to extend the gene body. This effectively makes the gene body larger as there
are proximal peaks that should be weighted equally to the gene body. This parameter is used if 'useTSS=FALSE'.}

\item{geneDownstream}{An integer describing the number of bp downstream the gene to extend the gene body.This effectively makes the gene body larger as there
are proximal peaks that should be weighted equally to the gene body. This parameter is used if 'useTSS=FALSE'.}

\item{useGeneBoundaries}{A boolean value indicating whether gene boundaries should be employed during gene activity score
calculation. Gene boundaries refers to the process of preventing tiles from contributing to the gene score of a given gene
if there is a second gene's transcription start site between the tile and the gene of interest.}

\item{useTSS}{A boolean describing whether to build gene model based on gene TSS or the gene body.}

\item{extendTSS}{A boolean describing whether to extend the gene TSS. By default useTSS uses the 1bp TSS while this parameter enables the extension of this
region with 'geneUpstream' and 'geneDownstream' respectively.}

\item{geneScaleFactor}{A numeric scaling factor to weight genes based on the inverse of there length i.e. \link{(Scale Factor)/(Gene Length)}. This
is scaled from 1 to the scale factor. Small genes will be the scale factor while extremely large genes will be closer to 1. This scaling helps with
the relative gene score value.}

\item{excludeChr}{A character vector containing the \code{seqnames} of the chromosomes that should be excluded from this analysis.}

\item{blacklist}{A \code{GRanges} object containing genomic regions to blacklist that may be extremeley over-represented and thus
biasing the geneScores for genes nearby that locus.}

\item{subThreads}{}

\item{tstart}{}

\item{logFile}{The path to a file to be used for logging ArchR output.}

\item{force}{A boolean value indicating whether to force the matrix indicated by \code{matrixName} to be overwritten if it already exist in the given \code{input}.}
}
\description{
This function, for each sample, will independently compute sums of score for each tile
per cell and then infer gene topic scores.
This function is adapted from \code{addGeneScoreMatrix} function from the  \code{ArchR} package.
}
