% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubblePlot.R
\name{bubblePlot}
\alias{bubblePlot}
\title{Title}
\usage{
bubblePlot(
  propMtrx = NULL,
  meanMtrx = NULL,
  markerSet = NULL,
  useFeatures = NULL,
  labelFeatures = useFeatures,
  Log2FC = 1,
  FDR = 0.05,
  row_title = "",
  column_title = "",
  filterSig = T,
  propCutoff = 0.1,
  color = "red3",
  ...
)
}
\arguments{
\item{row_title}{The same as \code{row_title} in \code{Heatmap} of \code{ComplexHeatmap}.}

\item{column_title}{The same as \code{column_title} in \code{Heatmap} of \code{ComplexHeatmap}.}

\item{...}{Other argument passed to \code{Heatmap} of \code{ComplexHeatmap}.}
}
\value{

}
\description{
Title
}
