% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchR_matrix.R
\name{getMeanMtrx}
\alias{getMeanMtrx}
\title{Get mean matrix for each group from an ArchR project}
\usage{
getMeanMtrx(
  proj = NULL,
  name = "GeneScoreMatrix",
  groupBy = "Clusters",
  useSeqnames = c("chr" \%&\% 1:22, "chrX"),
  threads = 4
)
}
\arguments{
\item{proj}{An ArchR project}

\item{name}{Name of matrix to use, default is GeneScoreMatrix}

\item{groupBy}{The variable in cellColData used to group cells, default is Clusters}

\item{useSeqnames}{Chromosomes to include in output matrix, default is chr1-22 and chrX}

\item{threads}{Number of threads to use, default is 4}
}
\value{
A feature-by-group sparse matrix
}
\description{
Get mean matrix for each group from an ArchR project
}
