% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchR_matrix.R
\name{getNonZeroProp}
\alias{getNonZeroProp}
\title{Get Non-zero proportion from ArchR project}
\usage{
getNonZeroProp(
  proj = NULL,
  name = "GeneScoreMatrix",
  groupBy = "Clusters",
  useSeqnames = c("chr" \%&\% 1:22, "chrX"),
  threads = 14
)
}
\arguments{
\item{proj}{An ArchR project}

\item{name}{Name of matrix to use, default is GeneScoreMatrix}

\item{groupBy}{The variable in cellColData used to group cells, default is Clusters}

\item{useSeqnames}{Chromosomes to include in output matrix, default is chr1-22 and chrX}

\item{threads}{Number of threads to use, default is 4}
}
\value{

}
\description{
Get Non-zero proportion from ArchR project
}
