% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchR_plot.R
\name{plotRedDim}
\alias{plotRedDim}
\title{Plot Reduced Dimensions}
\usage{
plotRedDim(mtrx, maxDim = 30, color = NULL, discrete = T, nCol = NULL)
}
\arguments{
\item{mtrx}{}

\item{maxDim}{The maximum number of dimensions to plot. This cannot be larger than the number of factors in mtrx}

\item{color}{A vector with equal length as row number of mtrx for coloring each point}

\item{discrete}{Whether the color should be discrete. Boolean. Default is TRUE}

\item{nCol}{}
}
\value{

}
\description{
Plot Reduced Dimensions
}
