% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchR_helper.R
\name{ArchR2sce}
\alias{ArchR2sce}
\title{Title}
\usage{
ArchR2sce(
  ArchRProj,
  matrix = "PeakMatrix",
  useSeqnames = c("chr" \%&\% 1:22, "chrX"),
  binarize = F,
  targetAssay = "count",
  scaleTo = 10000,
  reducedDims = c("IterativeLSI", "Harmony"),
  scaleDims = c(T, T),
  embeddings = "UMAP",
  useRowData = F,
  threads = 4
)
}
\arguments{
\item{ArchRProj}{}

\item{matrix}{}

\item{useSeqnames}{}

\item{binarize}{}

\item{targetAssay}{}

\item{scaleTo}{}

\item{reducedDims}{}

\item{scaleDims}{}

\item{embeddings}{}

\item{useRowData}{}

\item{threads}{}

\item{addLog}{}

\item{targetAssayLog}{}
}
\value{

}
\description{
Title
}
